requirements.txt
# explicit installs
torch==2.5.1
numpy==2.2.0
easydict==1.13
PyYAML==6.0.2
tqdm==4.67.1

# implicit installs (possible to comment out?)
filelock==3.16.1
fsspec==2024.10.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
sympy==1.13.1
typing_extensions==4.12.2